update omitted files for rename to pten
