+++ paddle/fluid/inference/CMakeLists.txt
@@ -35,7 +35,7 @@ endif()
 
 # fluid_modules exclude API-interface of inference/api and inference/capi_exp
 get_property(fluid_modules GLOBAL PROPERTY FLUID_MODULES)
-get_property(pten_modules GLOBAL PROPERTY TCMPT_MODULES)
+get_property(pten_modules GLOBAL PROPERTY PTEN_MODULES)
 
 # Adapt to custom op mechanism: Include the header files related to the data type
 # to avoid exposing the path of the underlying file
